modified orders page,bills page and payments page.. added extra information on the list (ex: description,glcd).
modified model pages of orders,bills,payments to add description.

+++ lib/models/customerdetails.dart
@@ -57,7 +57,7 @@ class RecordsetofCustomerDetails {
   double crBal;
   double dbBal;
   double opbal;
-  int crLimit;
+  double crLimit;
   String glcode;
 
   factory RecordsetofCustomerDetails.fromJson(Map<String, dynamic> json) => RecordsetofCustomerDetails(
@@ -74,7 +74,7 @@ class RecordsetofCustomerDetails {
     crBal: json["CR_BAL"].toDouble(),
     dbBal: json["DB_BAL"].toDouble(),
     opbal: json["OPBAL"].toDouble(),
-    crLimit: json["CR_LIMIT"],
+    crLimit: json["CR_LIMIT"].toDouble(),
     glcode: json["GLCODE"],
   );
